+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/subflows/index.ts
@@ -1,6 +1,3 @@
 export { LoopTool } from '@/app/workspace/[workspaceId]/w/[workflowId]/components/subflows/loop'
 export { ParallelTool } from '@/app/workspace/[workspaceId]/w/[workflowId]/components/subflows/parallel'
-export {
-  SubflowNodeComponent,
-  type SubflowNodeData,
-} from '@/app/workspace/[workspaceId]/w/[workflowId]/components/subflows/subflow-node'
+export { SubflowNodeComponent } from '@/app/workspace/[workspaceId]/w/[workflowId]/components/subflows/subflow-node'
