+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/utils/node-position-utils.ts
@@ -1,4 +1,4 @@
-import { BLOCK_DIMENSIONS, CONTAINER_DIMENSIONS } from '@/lib/workflows/blocks/block-dimensions'
+import { BLOCK_DIMENSIONS, CONTAINER_DIMENSIONS } from '@sim/workflow-renderer'
 import { getBlock } from '@/blocks/registry'
 
 /**
