+++ packages/workflow-renderer/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@sim/workflow-renderer",
+  "version": "0.1.0",
+  "private": true,
+  "sideEffects": [
+    "**/*.css",
+    "**/note-block-view.tsx"
+  ],
+  "type": "module",
+  "license": "Apache-2.0",
+  "engines": {
+    "bun": ">=1.2.13",
+    "node": ">=20.0.0"
+  },
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "default": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "type-check": "tsc --noEmit",
+    "lint": "biome check --write --unsafe .",
+    "lint:check": "biome check .",
+    "format": "biome format --write .",
+    "format:check": "biome format ."
+  },
+  "peerDependencies": {
+    "lucide-react": ">=0.479.0",
+    "react": "^19",
+    "reactflow": "^11.11.4",
+    "remark-breaks": "^4.0.0",
+    "streamdown": ">=2.5.0"
+  },
+  "devDependencies": {
+    "@sim/tsconfig": "workspace:*",
+    "@types/react": "^19",
+    "lucide-react": "^0.479.0",
+    "react": "19.2.4",
+    "reactflow": "^11.11.4",
+    "remark-breaks": "^4.0.0",
+    "streamdown": "2.5.0",
+    "typescript": "^5.7.3"
+  }
+}
