+++ packages/workflow-renderer/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@sim/tsconfig/library.json",
+  "compilerOptions": {
+    "jsx": "react-jsx",
+    "lib": ["ES2022", "DOM", "DOM.Iterable"]
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist"]
+}
